Rotary Dial View refactoring. Using a protocol.
